refactor: Update Logo component to support dynamic sizing
The Logo component has been updated to support dynamic sizing by adding a `size` prop. This allows the logo to be rendered at different sizes based on the provided value. The `sizeStyle` object is used to set the font size of the letter elements dynamically.

This change improves the flexibility and reusability of the Logo component, making it easier to customize the size of the logo based on the specific use case.