chore: Update next.config.mjs and postcss.config.mjs
This commit updates the next.config.mjs file to include a webpack configuration that sets up an alias for the "@" symbol to resolve to the current directory. It also sets the reactStrictMode flag to true. Additionally, the postcss.config.mjs file is updated to include the autoprefixer plugin. These changes improve the configuration and build process of the project.